@@ -2,18 +2,22 @@
   "name": "@next-book/publisher",
   "version": "0.2.24",
   "description": "Publish web books with next-book",
-  "main": "src/app.js",
+  "main": "dist/app.js",
   "bin": {
-    "publish-nb": "./bin/cli.js"
+    "publish-nb": "./dist/bin/cli.js"
   },
   "scripts": {
-    "bundle": "browserify src/browser.js -o dist/browser.js",
+    "tscRun": "tsc && node dist/app.js",
+    "dev": "onchange 'src/*' -v -- npm run tscRun",
+    "bundle": "browserify src/browser.ts -p [ tsify --noImplicitAny ] > dist/browser.js",
     "transpile": "babel dist/browser.js --out-file dist/browser.js",
     "updateExample": "cp dist/browser.js docs/browser.js",
-    "updateDocs": "npx documentation build src/**.js -f html -o ./docs/api --config documentation.yml --github",
-    "build": "npm run bundle && npm run transpile && npm run updateExample && npm run updateDocs",
+    "copyServiceWorkerTemplate": "cp src/service-worker/code.tpl.js dist/service-worker/code.tpl.js",
+    "updateDocs": "typedoc --out docs/api src/**.ts src/**/*.ts",
+    "build": "tsc && chmod u+x dist/bin/cli.js && npm run bundle && npm run transpile && npm run copyServiceWorkerTemplate && npm run updateExample && npm run updateDocs",
     "build:watch": "onchange 'src/*.js' -v -- npm run build",
-    "test": "jest"
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "clearJestCache": "jest --clearCache"
   },
   "repository": {
     "type": "git",
@@ -35,9 +39,8 @@
     "cli-progress": "^2.1.1",
     "commander": "^2.20.3",
     "express": "^4.17.1",
-    "i18next": "^19.3.2",
+    "i18next": "^20.3.5",
     "jsdom": "^16.6.0",
-    "object-hash": "^1.3.1",
     "pretty": "^2.0.0",
     "recursive-copy": "^2.0.10",
     "recursive-readdir": "^2.2.2",
@@ -51,14 +54,28 @@
     "@babel/cli": "^7.12.1",
     "@babel/core": "^7.7.5",
     "@babel/preset-env": "^7.7.6",
+    "@babel/preset-typescript": "^7.15.0",
+    "@types/cli-progress": "^3.9.2",
+    "@types/express": "^4.17.13",
+    "@types/jest": "^26.0.24",
+    "@types/jsdom": "^16.2.13",
+    "@types/node": "^16.4.13",
+    "@types/pretty": "^2.0.0",
+    "@types/recursive-readdir": "^2.2.0",
+    "@types/rimraf": "^3.0.1",
+    "@types/slug": "^5.0.2",
+    "@typescript-eslint/eslint-plugin": "^4.29.0",
+    "@typescript-eslint/parser": "^4.29.0",
     "browserify": "^16.5.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.1",
-    "eslint-config-prettier": "^3.6.0",
-    "eslint-plugin-import": "^2.19.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.17.0",
-    "jest": "^25.0.0",
-    "onchange": "^6.1.0"
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.24.0",
+    "eslint-plugin-tsdoc": "^0.2.14",
+    "jest": "^27.0.6",
+    "onchange": "^6.1.0",
+    "ts-jest": "^27.0.4",
+    "tsify": "^5.0.4",
+    "typedoc": "^0.21.5",
+    "typescript": "^4.3.5"
   }
 }
