@@ -1,6 +1,7 @@
-const Jsdom = require('jsdom').JSDOM;
-const parser = require('./parser.js');
-const { ParsedObj } = require('./structures.js');
+/* eslint-disable no-irregular-whitespace */
+import { ParsedObj } from './structures';
+import { JSDOM as Jsdom } from 'jsdom';
+import parse from './parser';
 
 const dom = new Jsdom(`<main><p><q>Vidíš,</q> radoval se trochu nuceně, <q>je to náhoda!
 Mně totiž dnes ráno napadlo, že už jsme se dvacet let neviděli!
@@ -23,8 +24,7 @@ Honil je sloupec za sloupcem; ukrylo se s rozčilující schválností.
 Nyní pan Bondy začal zdola a konečně od pravé strany.
 Protivný NÁLEZ byl ten tam.</p></main>`);
 
-test('parser works', () => {
-  expect(parser.parse(dom.window.document.querySelector('main'), '\n') instanceof ParsedObj).toBe(
-    true
-  );
+test('Parser returns ParsedObj', () => {
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  expect(parse(dom.window.document.querySelector('main')!, '\n') instanceof ParsedObj).toBe(true);
 });
